+++ Source/SwiftLintBuiltInRules/Rules/RuleConfigurations/PrivateUnitTestConfiguration.swift
@@ -1,49 +1,56 @@
 import Foundation
 import SwiftLintCore
 
-struct PrivateUnitTestConfiguration: SeverityBasedRuleConfiguration, Equatable, CacheDescriptionProvider {
+struct PrivateUnitTestConfiguration: SeverityBasedRuleConfiguration, Equatable {
     typealias Parent = PrivateUnitTestRule
 
     @ConfigurationElement(key: "severity")
     private(set) var severityConfiguration = SeverityConfiguration<Parent>(.warning)
+    @ConfigurationElement(key: "test_parent_classes")
+    private(set) var testParentClasses: Set<String> = ["QuickSpec", "XCTestCase"]
+
     @ConfigurationElement(key: "regex")
     private(set) var regex = SwiftLintCore.regex("XCTestCase")
-
-    private(set) var name: String?
-    private(set) var message = "Unit test marked `private` will not be run by XCTest."
-    private(set) var included: NSRegularExpression?
-
-    var cacheDescription: String {
-        let jsonObject: [String] = [
-            "private_unit_test",
-            name ?? "",
-            message,
-            regex.pattern,
-            included?.pattern ?? "",
-            severityConfiguration.severity.rawValue
-        ]
-        if let jsonData = try? JSONSerialization.data(withJSONObject: jsonObject),
-          let jsonString = String(data: jsonData, encoding: .utf8) {
-              return jsonString
-        }
-        queuedFatalError("Could not serialize private unit test configuration for cache")
-    }
 
     mutating func apply(configuration: Any) throws {
         guard let configurationDict = configuration as? [String: Any] else {
             throw Issue.unknownConfiguration(ruleID: Parent.identifier)
         }
+        if let extraTestParentClasses = configurationDict[$testParentClasses] as? [String] {
+            self.testParentClasses.formUnion(extraTestParentClasses)
+        }
         if let regexString = configurationDict[$regex] as? String {
+            // TODO: [01/09/2025] Remove deprecation warning after ~2 years and use `UnitTestConfiguration`
+            // instead of this configuration.
+            queuedPrintError(
+                """
+                warning: '\($regex)' has been replaced by a list of explicit parent class names. They can be \
+                configured in the '\($testParentClasses)' option. '\($regex)' will be completely removed \
+                in a future release.
+                """
+            )
             regex = try .cached(pattern: regexString)
         }
-        if let includedString = configurationDict["included"] as? String {
-            included = try .cached(pattern: includedString)
+        if configurationDict["included"] is String {
+            // TODO: [01/09/2025] Remove deprecation warning after ~2 years and replace this configuration by
+            // `UnitTestConfiguration`.
+            queuedPrintError(
+                "warning: 'included' is ignored from now on. You may remove it from the configuration file."
+            )
         }
-        if let name = configurationDict["name"] as? String {
-            self.name = name
+        if configurationDict["name"] is String {
+            // TODO: [01/09/2025] Remove deprecation warning after ~2 years and replace this configuration by
+            // `UnitTestConfiguration`.
+            queuedPrintError(
+                "warning: 'name' is ignored from now on. You may remove it from the configuration file."
+            )
         }
-        if let message = configurationDict["message"] as? String {
-            self.message = message
+        if configurationDict["message"] is String {
+            // TODO: [01/09/2025] Remove deprecation warning after ~2 years and replace this configuration by
+            // `UnitTestConfiguration`.
+            queuedPrintError(
+                "warning: 'message' is ignored from now on. You may remove it from the configuration file."
+            )
         }
         if let severityString = configurationDict[$severityConfiguration] as? String {
             try severityConfiguration.apply(configuration: severityString)